@@ -1,42 +1,56 @@
 {
-  "name": "fyle-frontend-challenge",
+  "name": "application2",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test"
+    "test": "ng test",
+    "serve:ssr:Application2": "node dist/application2/server/server.mjs"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.1.0",
-    "@angular/common": "^16.1.0",
-    "@angular/compiler": "^16.1.0",
-    "@angular/core": "^16.1.0",
-    "@angular/forms": "^16.1.0",
-    "@angular/platform-browser": "^16.1.0",
-    "@angular/platform-browser-dynamic": "^16.1.0",
-    "@angular/router": "^16.1.0",
-    "octokit": "^3.1.1",
+    "@angular/animations": "^17.3.11",
+    "@angular/cdk": "^17.3.10",
+    "@angular/common": "^17.2.0",
+    "@angular/compiler": "^17.2.0",
+    "@angular/core": "^17.2.0",
+    "@angular/forms": "^17.2.0",
+    "@angular/material": "^17.3.10",
+    "@angular/platform-browser": "^17.2.0",
+    "@angular/platform-browser-dynamic": "^17.2.0",
+    "@angular/platform-server": "^17.2.0",
+    "@angular/router": "^17.2.0",
+    "@angular/ssr": "^17.2.2",
+    "@fortawesome/fontawesome-free": "^6.5.2",
+    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+    "@popperjs/core": "^2.11.8",
+    "bootstrap": "^5.3.3",
+    "chart.js": "^3.9.1",
+    "express": "^4.18.2",
+    "ng2-charts": "^3.1.2",
+    "primeicons": "^7.0.0",
+    "primeng": "^17.18.4",
     "rxjs": "~7.8.0",
+    "tailwindcss": "^3.4.4",
     "tslib": "^2.3.0",
-    "zone.js": "~0.13.0"
+    "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.1.8",
-    "@angular/cli": "~16.1.8",
-    "@angular/compiler-cli": "^16.1.0",
-    "@types/jasmine": "~4.3.0",
-    "autoprefixer": "^10.4.16",
-    "jasmine-core": "~4.6.0",
+    "@angular-devkit/build-angular": "^17.2.2",
+    "@angular/cli": "^17.2.2",
+    "@angular/compiler-cli": "^17.2.0",
+    "@angular/localize": "^17.2.0",
+    "@types/express": "^4.17.17",
+    "@types/jasmine": "~5.1.0",
+    "@types/node": "^18.18.0",
+    "jasmine-core": "~5.1.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "postcss": "^8.4.31",
-    "tailwindcss": "^3.3.5",
-    "typescript": "~5.1.3"
+    "typescript": "~5.3.2"
   }
 }
