@@ -1,17 +1,10 @@
-import { Component, OnInit } from '@angular/core';
-import { ApiService } from './services/api.service';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+  styleUrl: './app.component.css'
 })
-export class AppComponent implements OnInit{
-  constructor(
-    private apiService: ApiService
-  ) {}
-
-  ngOnInit() {
-    this.apiService.getUser('johnpapa').subscribe(console.log);
-  }
+export class AppComponent {
+  title = 'Application2';
 }
